-[Feature]: Multi-line inputs for adding tracker URLs
- moved classes EditTextView and EditTextViewController into CustomViews as swift files
