+++ iTorrent/Utils/Dialog.swift
@@ -109,191 +109,3 @@ class Dialog {
         return nil
     }
 }
-
-
-// EditTextView - UI Control
-class EditTextView: UITextView{
-    
-    // Set margins using UIEdgeInsets
-    let defaultMargins      = UIEdgeInsets(top: 4, left: 12, bottom: 12, right: 12)
-    let defaultPlaceholder  = NSLocalizedString("PlaceHolder text", comment: "")
-    
-    // Custom placeholder text
-    var placeholder: String? {
-        didSet { placeholderLabel.text = placeholder }
-    }
-    
-    // Custom placeholder text color
-    var placeholderTextColor: UIColor = .lightGray {
-        didSet { placeholderLabel.textColor = placeholderTextColor }
-    }
-    
-    // Private UILabel for the placeholder text
-    private let placeholderLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = .lightGray
-        return label
-    }()
-
-    
-    /* Override methods */
-    override init(frame: CGRect, textContainer: NSTextContainer?){
-        super.init(frame: frame, textContainer:textContainer)
-        performCustomizations()
-    }
-    
-    required init?(coder: NSCoder){
-        super.init(coder:coder)
-        performCustomizations()
-    }
-    
-    
-    /* Helper methods */
-    private func performCustomizations() -> Void{
-        let textView = self
-        let theme = Themes.current
-        
-        textView.keyboardAppearance = theme.keyboardAppearence
-        textView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-        textView.font = UIFont.systemFont(ofSize: UIFont.systemFontSize * 1.0)
-        textView.contentInset.left      = 2         //For left padding
-        textView.contentInset.right     = 2         //For right padding
-        textView.contentInset.top       = 2         //For top padding
-        textView.contentInset.bottom    = 2         //For bottom padding
-        
-        textView.layer.borderWidth = 0.5
-        textView.layer.cornerRadius = 8
-        textView.layer.borderColor = UIColor.lightGray.withAlphaComponent(0.5).cgColor
-        textView.layer.backgroundColor = nil
-        textView.returnKeyType = .default
-        textView.autocapitalizationType = .none
-       
-        textView.translatesAutoresizingMaskIntoConstraints = false
-        setupPlaceholderLabel()
-    }
-    
-    private func setupPlaceholderLabel() {
-        addSubview(placeholderLabel)
-        
-        placeholder = self.defaultPlaceholder
-        placeholderLabel.font = UIFont.systemFont(ofSize: UIFont.systemFontSize * 1.0)
-        placeholderLabel.layer.borderColor = UIColor.lightGray.withAlphaComponent(0.5).cgColor
-
-        placeholderLabel.translatesAutoresizingMaskIntoConstraints = false
-        // Set constraints for the placeholder label
-        NSLayoutConstraint.activate([
-            placeholderLabel.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: self.contentInset.left+4),
-            placeholderLabel.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -(self.contentInset.right+4)),
-            placeholderLabel.topAnchor.constraint(equalTo: self.topAnchor, constant: self.contentInset.top+6),
-            placeholderLabel.bottomAnchor.constraint(equalTo: self.bottomAnchor, constant: -(self.contentInset.bottom+6))
-        ])
-        
-        // Hide the placeholder label when there is text
-        NotificationCenter.default.addObserver(self, selector: #selector(textDidChange), name: UITextView.textDidChangeNotification, object: nil)
-    }
-    
-    @objc private func textDidChange() {
-        placeholderLabel.isHidden = !text.isEmpty
-    }
-    
-    public func setDefaultMargins(){
-        setMargins(by: self.defaultMargins)
-    }
-    
-    
-    public func setMargins(by: UIEdgeInsets){
-        let margins = by
-        if let containerView = self.superview{
-            self.frame = containerView.frame
-            // Add constraints to the UITextView within the parent view
-            NSLayoutConstraint.activate([
-                self.leadingAnchor.constraint(equalTo: containerView.leadingAnchor, constant: margins.left),
-                self.trailingAnchor.constraint(equalTo: containerView.trailingAnchor, constant: -margins.right),
-                self.topAnchor.constraint(equalTo: containerView.topAnchor, constant: margins.top),
-                self.bottomAnchor.constraint(equalTo: containerView.bottomAnchor, constant: -margins.bottom)
-            ])
-        }
-    }
-    
-    // callback for superview changes
-    override func didMoveToSuperview() {
-        // attach superView's frame reference into this view
-        // and set margins if superView is present
-        self.setDefaultMargins()
-    }
-}
-
-
-class EditTextViewController: UIViewController, UITextViewDelegate {
-    
-    private let editTextView:EditTextView
-    private let SINGLE_LINE_HEIGHT:Double = 54.0
-
-    convenience init(){
-        self.init(editTextView: EditTextView())
-    }
-
-    init(editTextView: EditTextView){
-        self.editTextView = editTextView
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    required init?(coder: NSCoder) {
-        self.editTextView = EditTextView()
-        super.init(coder: coder)
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // add the UITextView
-        self.view.addSubview(self.editTextView)
-        // register callback for textView changes
-        editTextView.delegate = self
-        // Create a height constraint for the view controller's view
-        viewControllerHeightConstraint = view.heightAnchor.constraint(equalToConstant: SINGLE_LINE_HEIGHT)
-        viewControllerHeightConstraint.isActive = true
-    }
-    
-    var viewControllerHeightConstraint: NSLayoutConstraint!
-    
-    func textViewDidChange(_ textView: UITextView) {
-
-        let maxLinesToShow:Double = 1.5
-        let maxContentSize:Double = SINGLE_LINE_HEIGHT * maxLinesToShow
-        
-        var contentSize:Double = textView.contentSize.height + 17.0
-        if contentSize > maxContentSize{
-            contentSize = maxContentSize
-        }
-        if contentSize < SINGLE_LINE_HEIGHT{
-            contentSize = SINGLE_LINE_HEIGHT
-        }
-        
-        if(contentSize != prevContentSize){
-            // Calculate the content height of the UITextView
-            // Update the height constraint of the view controller's view
-            viewControllerHeightConstraint.constant = contentSize
-
-            // Optionally, animate the constraint change
-            UIView.animate(
-                       withDuration: 0.5,
-                       delay: 0.0,
-                       usingSpringWithDamping: 0.5, // Adjust this value (lower for more bounce, higher for less)
-                       initialSpringVelocity: 0.0,  //Adjust this value (higher for more initial velocity)
-                       options: [.curveEaseInOut],  // Use curveEaseInOut
-                       animations: {
-                          self.view.layoutIfNeeded()
-                       },
-                       completion: nil
-            )
-        }
-        prevContentSize = contentSize
-        
-        
-        // Optionally, you can scroll to the bottom as the user types
-//        let bottomOffset = CGPoint(x: 0, y: max(textView.contentSize.height - textView.bounds.size.height, 0))
-//        textView.setContentOffset(bottomOffset, animated: false)
-    }
-    
-    private var prevContentSize: Double = 0.0
-}
